Made utility class static methods
